requirements.txt
addict==2.2.0
anykeystore==0.2
apex==0.9.10.dev0
batchgenerators==0.19.3
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptacular==1.5.5
cycler==0.10.0
Cython==0.29.6
DCN==1.0
decorator==4.4.0
defusedxml==0.6.0
dicom2nifti==2.1.1
dill==0.3.0
fast-fcn==0.5.1+097e713
future==0.17.1
hupper==1.9.1
idna==2.8
imageio==2.5.0
kiwisolver==1.0.1
linecache2==1.0.0
matplotlib==3.0.3
MedPy==0.4.0
mkl-fft==1.0.10
mkl-random==1.0.2
mmcv==0.2.6
-e git+https://github.com/open-mmlab/mmdetection.git@a3c8ddf766d76693013a16e40f79dd4fedf2db7f#egg=mmdet
networkx==2.3
nibabel==2.4.0
nose==1.3.7
numpy==1.16.2
oauthlib==3.1.0
opencv-python==4.0.0.21
pandas==0.24.2
PasteDeploy==2.0.1
pbkdf2==1.3
Pillow==5.4.1
plaster==1.0
plaster-pastedeploy==0.7
protobuf==3.7.1
pycocotools==2.0.0
pycparser==2.19
pydicom==1.2.2
pyparsing==2.3.1
pyramid==1.10.4
pyramid-mailer==0.15.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1
repoze.sendmail==4.4.1
requests==2.21.0
requests-oauthlib==1.3.0
scikit-image==0.15.0
scikit-learn==0.20.3
scipy==1.2.1
SimpleITK==1.2.0
six==1.12.0
sklearn==0.0
SQLAlchemy==1.3.12
tensorboardX==1.6
terminaltables==3.1.0
thop==0.0.22
torch==1.0.1.post2
torchvision==0.2.2.post3
tqdm==4.31.1
traceback2==1.4.0
transaction==3.0.0
translationstring==1.3
typing==3.6.4
unittest2==1.1.0
urllib3==1.24.2
velruse==1.1.1
venusian==3.0.0
WebOb==1.8.5
WTForms==2.2.1
wtforms-recaptcha==0.3.2
zope.deprecation==4.4.0
zope.interface==4.7.1
zope.sqlalchemy==1.2
